Rename example project. Create SDK project file.  Add additional files.
