+++ .swift-version
@@ -1,0 +1,1 @@
+4.2
