+++ Example/Example/AppDelegate.swift
@@ -10,6 +10,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
         // On first start of the app call this:
         InPlayer.initialize(withClienId: "xxx", environment: .staging)
+
         return true
     }
 }
