+++ Podfile
@@ -1,0 +1,13 @@
+platform :ios, '10.0'
+
+target 'InPlayerSDK' do
+  use_frameworks!
+
+  pod 'Alamofire', '5.0.0.beta.1'
+  pod 'AWSIoT', '2.8.4'
+
+target 'InPlayerSDKTests' do
+    inherit! :search_paths
+  end
+
+end
