+++ Source/Notification/Notification.swift
@@ -5,6 +5,14 @@ import AWSCore
 private protocol AWSNotification {
 
     /**
+     Connects to a websocket and subscribes for `InPlayerNotification`s.
+     - Parameters:
+        - statusCallback: A closure to be executed everytime when connection status change occurs.
+        - status: `InPlayerNotificationStatus` enum that shows current connection status
+        - messageCallback: A closure to be executed everytime new notification is sent from the server.
+        - notification: `InPlayerNotification` enum optionally containing associated values depending on the type.
+        - onError: A closure to be executed when some error occurred.
+        - error: `InPlayerError` containing info about the error that occurred.
      */
     static func subscribe(statusCallback: @escaping (_ status: InPlayerNotificationStatus) -> Void,
                           messageCallback: @escaping (_ notification: InPlayerNotification) -> Void,
@@ -12,6 +20,7 @@ private protocol AWSNotification {
 
 
     /**
+     Disconnects from websocket and stop receving `InPlayerNotification`.
      */
     static func disconnect()
 }
@@ -23,13 +32,13 @@ public extension InPlayer {
         public static func subscribe(statusCallback: @escaping (_ status: InPlayerNotificationStatus) -> Void,
                                      messageCallback: @escaping (_ notification: InPlayerNotification) -> Void,
                                      onError: @escaping (_ error: InPlayerError) -> Void) {
-            INPNotificationManager.subscribe(statusCallback: statusCallback,
-                                             messageCallback: messageCallback,
-                                             onError: onError)
+            INPAWSManager.subscribe(statusCallback: statusCallback,
+                                    messageCallback: messageCallback,
+                                    onError: onError)
         }
 
         public static func disconnect() {
-            INPNotificationManager.disconnect()
+            INPAWSManager.disconnect()
         }
     }
 }
